move Eth2NetworkOptions converters to their own classes
