+++ teku/src/main/java/tech/pegasys/teku/cli/converter/UInt256Converter.java
@@ -1,0 +1,30 @@
+/*
+ * Copyright 2021 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.cli.converter;
+
+import java.math.BigInteger;
+import org.apache.tuweni.units.bigints.UInt256;
+import picocli.CommandLine.ITypeConverter;
+import picocli.CommandLine.TypeConversionException;
+
+public class UInt256Converter implements ITypeConverter<UInt256> {
+  @Override
+  public UInt256 convert(final String value) {
+    try {
+      return UInt256.valueOf(new BigInteger(value));
+    } catch (final NumberFormatException e) {
+      throw new TypeConversionException("Invalid format: must be a numeric value but was " + value);
+    }
+  }
+}
